add authentication service and refactor login logic in AuthController
