+++ interfaces/IAuthService.cs
@@ -1,0 +1,8 @@
+using vueChain.Dtos;
+namespace vueChain.Interfaces
+{
+    public interface IAuthService
+    {
+        Task<string> Login(UserDto userDto);
+    }
+}
