+++ services/AuthService.cs
@@ -1,0 +1,55 @@
+using vueChain.Interfaces;
+using vueChain.Dtos;
+using vueChain.Models;
+using Microsoft.IdentityModel.Tokens;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
+
+namespace vueChain.Services
+{
+    public class AuthService : IAuthService
+    {
+        private readonly IUserService _userService;
+        private readonly IConfiguration _configuration;
+
+        public AuthService(IUserService userService, IConfiguration configuration)
+        {
+            _userService = userService;
+            _configuration = configuration;
+        }
+
+        public async Task<string> Login(UserDto userDto)
+        {
+            var user = await _userService.GetUserByUsername(userDto.Username);
+            if (user == null || !VerifyPasswordHash(userDto.Password, user.PasswordHash))
+            {
+                return null;
+            }
+
+            return GenerateJwtToken(user);
+        }
+
+        private bool VerifyPasswordHash(string password, string storedHash)
+        {
+            return BCrypt.Net.BCrypt.Verify(password, storedHash);
+        }
+
+        private string GenerateJwtToken(User user)
+        {
+            var tokenHandler = new JwtSecurityTokenHandler();
+            var key = Encoding.ASCII.GetBytes(_configuration["Jwt:Key"]);
+            var tokenDescriptor = new SecurityTokenDescriptor
+            {
+                Subject = new ClaimsIdentity(new Claim[]
+                {
+                    new Claim(ClaimTypes.Name, user.Username)
+                }),
+                Expires = DateTime.UtcNow.AddDays(7),
+                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
+            };
+            var token = tokenHandler.CreateToken(tokenDescriptor);
+            return tokenHandler.WriteToken(token);
+        }
+    }
+}
